inner algorithm completed

+++ main/views.py
@@ -39,7 +39,7 @@ def message(request):
             
         else:
             
-            r_s_list=[x.lecture_name for x in user.recent_search_set.all()]
+            r_s_list=["@"+x.lecture_name for x in user.recent_search_set.all()]
             
             response_json={
                 "message":{
@@ -62,18 +62,61 @@ def message(request):
             }
         }
     
+    #---강의 이름이 들어왔을 경우
     else:
-        if ":" in content:
+        
+        #---맨 앞 @가 붙은 경우에는 최근 검색 결과에서 검색한 경우
+        if "@" is content[0]:
+          
+            search_content=content[1:]
+            recent=user.recent_search.get(lecture_name=search_content)
+            lecture_list=lecture.objects.filter(
+                lecture_name=recent.lecture_name,
+                professor_name=recent.professor_name,
+                course_number=recent.course_number)
+        
+        #---강의 이름만으로 검색 했을 때 여러개의 강의가 나온 경우 :로 강의이름 교수 학수번호를 구분해 세부검색 
+        elif ":" in content:
             
             temp=content.split(":")
             lecture_list=lecture.objects.filter(lecture_name=temp[0],professor_name=temp[1],course_number=temp[2])
-            
+        
+        #---아예 처음 검색한 경우
         else:
             
             lecture_list=lecture.objects.filter(lecture_name__icontains=content)
+
+
+        #---검색 했을 때 결과가 0인 경우
+        if lecture_list.count()==0:
         
-        
-        if lecture_list.count()>1:
+            if user.recent_search_set.all().count()==0:
+            
+                response_json={
+                    "message":{
+                        "text": "해당 강의가 없습니다! 강의 이름을 다시 입력해 주세요"
+                    },
+                    "keyboard":{
+                        "type": "text"
+                    }
+                }
+                
+            else:
+                
+                r_s_list=["@"+x.lecture_name for x in user.recent_search_set.all()]
+                
+                response_json={
+                    "message":{
+                        "text": "해당 강의가 없습니다! 최근 검색한 강의 중 선택하거나<br>강의 검색을 눌러주세요"
+                    },
+                    "keyboard":{
+                        "type": "buttons",
+                        "buttons":["강의 검색"]+r_s_list
+                    }
+                }
+
+        #---검색 했을 때 결과가 여러개 나온 경우 -> :를 포함한 세부 검색 결과로 진행
+        elif lecture_list.count()>1:
             
             lecture_list=[lecture.lecture_name+":"+lecture.professor_name+":"+lecture.course_number for lecture in lecture_list]
             
@@ -86,20 +129,39 @@ def message(request):
                     "buttons": lecture_list
                 }
             }
-            
+        
+        #---검색 했을 때 결과가 한 개인 경우(검색에 성공한 경우)
         else:
             
-            if user.recent_search_set.all().count()>5:
-                
-                user.recent_search_set.all()[0].delete()
-                recent_search(kakao_user=user,lecture_name=temp[0],professor_name=temp[1],course_number=temp[2]).save()
-                
-            else:
-                
-                recent_search(kakao_user=user,lecture_name=temp[0],professor_name=temp[1],course_number=temp[2]).save()
-            #검색 후 보여주기
+            r_s_list=["@"+x.lecture_name for x in user.recent_search_set.all()]
             
-        
+            #---최근 검색 강의에 먼저 저장
+            if user.recent_search_set.all().count()>5: user.recent_search_set.all()[0].delete()
+            
+            recent_search(kakao_user=user,
+                    lecture_name=lecture_list[0].lecture_name,
+                    professor_name=lecture_list[0].professor_name,
+                    course_number=lecture_list[0].course_name).save()
+            
+            lecture_list[0].popularity+=1
+            lecture_list[0].save()
+            
+            response_json={
+                "message":{
+                    "text": lecture_list[0].lecture_name+"<br>"+
+                        lecture_list[0].professor_name+"<br>"+
+                        str(lecture_list[0].opening)+"/"+str(lecture_list[0].total_number)
+                },
+                "keyboard":{
+                    "type": "buttons",
+                    "buttons": ["강의 검색"]+r_s_list
+                }
+            }
+            if lecture_list[0].opening < lecture_list[0].total_number:
+                response_json["message"]["message_button"]={"label": "자리남!!!", "url": "http://www.hufs.ac.kr"}
+
+    return HttpResponse(json.dumps(response_json,ensure_ascii=False), content_type=u"application/json; charset=utf-8")
+
 @csrf_exempt    
 def reg_friend(request):
     value=json.loads(request.body.decode("utf-8"))
